Invariant hierarchy with schemas and mutability.
A hierarchy of invariant levels, indexed by natural numbers, is introduced.
Standard Iris invariants are at the lowest level. The fancy update modality
also has a version which is annotated with a level, which only allows
invariants up to that level to be opened. The standard unannotated fupd can
open any level.

A more general form of invariants with schemas is added. These allow encoding
assertions in the logic from a fixed grammar, with holes into which arbitrary
propositions can be inserted. When opening an invariant, only the assertions at
the holes are guarded by laters. A schema can refer to fupd at lower levels in
the hierarchy, allowing nesting of invariants without laters, as long as there
is stratification of the levels. The usual impredicative use of invariants can
still be recovered by inserting an invariant into a hole.

Finally, we also have mutable holes in invariant schemas. A thread can change
the proposition that goes into that hole with a non-duplicable, fractional
permission. If the invariant is opened without that permission, the
corresponding propositions for those holes are existentially quantified.